requirements.txt
drf-yasg
djangorestframework-simplejwt
django-filter
asgiref
Django
mysqlclient
PyMySQL
python-dotenv
sqlparse

djangorestframework
django_filters